+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Include `python_api.h` for TensorFlow ([issue #585](https://github.com/bytedeco/javacpp-presets/issues/585))
  * Add presets for ONNX 1.2.2 ([pull #547](https://github.com/bytedeco/javacpp-presets/pull/547))
  * Define `std::vector<tensorflow::OpDef>` type to `OpDefVector` for TensorFlow
  * Link HDF5 with zlib on Windows also ([issue deeplearning4j/deeplearning4j#6017](https://github.com/deeplearning4j/deeplearning4j/issues/6017))